requirements.txt
langchain
unstructured # Document loading
openai # For embeddings
tiktoken  # For embeddings
beautifulsoup4
nltk
chardet
langchain-openai
pinecone-client
langchain-pinecone
streamlit
pdfminer.six
python-dotenv
